Limit the byte size of the bulk update (#1)
* Limit the byte size of the bulk update

Instead of triggering a flush of the bulk based on a hard coded limit of
1024 documents which could represents a very large range of sizes,
trigger based on the byte size of the batch of documents. This means
we'll get larger batches for very small documents which may improve
throughput, but more importantly it fixes a bug when our batch of
documents exceeds a hard-limit in mongodb server.
This limits the size to below the recommended maximum size of 16M.

* Remember to reset the byte counter

Reset the counter otherwise, we flush very often after the initial flush
happens.